scrapers/requirements.txt
requests>=2.32.3
pymongo>=4.10.1
beautifulsoup4==4.12.3
lxml>=5.0.0
python-dotenv==1.0.1
playwright>=1.51.0
